pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "demapi"
version = "0.1.0"
description = "Make customizable demotivators and motivators through imgonline.com.ua API. Supports async-await style"
authors = ["Yan Kurbatov <deknowny@gmail.com>"]
license = "GNU General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
aiohttp = "^3.8.1"
certifi = "^2021.10.8"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
isort = "^5.10.1"
mypy = "^0.910"
pre-commit = "^2.15.0"
pytest-cov = "^3.0.0"
coveralls = "^3.3.1"
coverage = "^6.1.2"
pytest-asyncio = "^0.16.0"

[tool.black]
line-length = 78

[tool.isort]
profile = "wemake"
src_paths = ["demapi", "tests"]
line_length = 78
